♻️ Refactor: 메인 페이지 컴포넌트화 - 서버, 클라이언트 나눠서 컴포넌트화 진행 - 기존의 메인페이지파일을 app 의 하위 폴더로 이동 - interface와 같은 타입을 정의하기 위한 코드들을 types/ 의 하위 폴더로 이동 Close #107
