+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import MainPage from '@/pages/mainPage';
+import MainPage from '@/app/mainPage/page';
 
 export default function Home() {
   return <MainPage />;
