+++ src/styles/mainPage/sections/eventSection.module.css
@@ -1,0 +1,54 @@
+@import url(../../globals.css);
+@import url(../../reset/modern-reset.css);
+
+/* 이벤트 팝업 영역  */
+
+.event_wrapper {
+  width: 100%;
+  position: relative; /* 기준점 */
+  aspect-ratio: clamp(500/270, 760/360, 870/570);
+  display: block; /* 흐름에 참여하도록 */
+}
+
+.event_wrapper .img {
+  position: relative;
+  width: 100%;
+  height: 100%;
+}
+
+.event_popup {
+  position: absolute; /* 이 부분이 가운데 정렬 */
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  z-index: 2;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 100%;
+  gap: clamp(6px, 2vw, 16px);
+}
+
+.main_image_wrapper {
+  position: relative;
+  flex: 1;
+  width: 100%;
+  padding: 13px;
+  max-width: clamp(55%, 65vw, 65%);
+  aspect-ratio: 790 / 200;
+}
+
+.sub_event_popup {
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  gap: clamp(12px, 2vw, 15px);
+  width: 100%;
+  max-width: clamp(55%, 55vw, 65%);
+}
+
+.sub_image_wrapper {
+  position: relative;
+  flex: 1;
+  aspect-ratio: 288 / 97;
+}
