+++ src/styles/mainPage/sections/productSection.module.css
@@ -1,0 +1,225 @@
+@import url(../../globals.css);
+@import url(../../reset/modern-reset.css);
+
+/* 메인페이지 상품 영역 */
+.product_section {
+  margin-bottom: 40px;
+}
+.product_rows {
+  display: flex;
+  flex-direction: column;
+}
+
+.product_wrapper {
+  display: flex;
+  flex-direction: row;
+  height: fit-content;
+  border-block: 1px solid #b6c5c5;
+}
+
+@media (max-width: 900px) {
+  .product_wrapper {
+    padding-block: 22px;
+    justify-content: center;
+  }
+}
+
+.product_hero {
+  flex: 0 0 360px;
+}
+
+@media (max-width: 900px) {
+  .product_hero {
+    display: none;
+    width: 0;
+    flex: 0 0 0;
+    margin: 0;
+    padding: 0;
+  }
+}
+
+.product_row {
+  width: 100%;
+  overflow: hidden;
+}
+
+.product_row_scroll {
+  display: flex;
+  flex: 1;
+  gap: 10px;
+  overflow-x: auto;
+  padding: 10px 0;
+  padding-bottom: 20px;
+  scroll-behavior: smooth;
+  align-self: center;
+  margin-left: 100px;
+  scrollbar-width: thin;
+  scrollbar-color: #081818 #b6c5c5;
+  -webkit-overflow-scrolling: touch;
+
+  @media (min-width: 900px) {
+    padding-bottom: 50px;
+    gap: 30px;
+  }
+}
+
+@media (max-width: 900px) {
+  .product_row_scroll {
+    margin-left: 0;
+  }
+}
+
+.product_row:nth-child(even) .product_wrapper {
+  flex-direction: row-reverse;
+  border: 0px;
+}
+
+.product_row:nth-child(even) .product_row_scroll {
+  margin-right: 20px;
+
+  @media (min-width: 1200px) {
+    margin-left: 250px;
+  }
+}
+
+@media (max-width: 900px) {
+  .product_row:nth-child(even) .product_row_scroll {
+    margin-right: 0;
+  }
+}
+
+.product_card {
+  overflow: hidden;
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  cursor: pointer;
+  width: 140px;
+  height: 270px;
+  flex-shrink: 0;
+
+  @media (min-width: 900px) {
+    width: 180px;
+    height: 335px;
+  }
+}
+
+.product_image {
+  height: 100%;
+  width: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  position: relative;
+  overflow: hidden;
+  margin-bottom: 15px;
+
+  @media (min-width: 900px) {
+    margin-bottom: 30px;
+  }
+}
+
+.product_image img {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+}
+
+.product_info {
+  width: 100%;
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+}
+
+.product_category {
+  font-size: 14px;
+  font-weight: 600;
+  color: var(--color-text);
+  margin-bottom: 5px;
+  font-weight: 700;
+
+  @media (min-width: 900px) {
+    font-size: 16px;
+  }
+}
+
+.product_name {
+  font-size: 14px;
+  font-weight: 700;
+  color: var(--color-text);
+  margin-bottom: 6px;
+  line-height: 1.3;
+  overflow: hidden;
+  flex: 1;
+  white-space: nowrap;
+  text-overflow: ellipsis;
+
+  @media (min-width: 900px) {
+    font-size: 16px;
+    margin-bottom: 10px;
+  }
+}
+
+.product_price {
+  font-size: 15px;
+  font-weight: bold;
+  color: var(--color-text-pointer);
+  margin-top: auto;
+}
+
+/* 카테고리 영역 */
+.section_header {
+  margin-block: 29px;
+  margin-inline: 16px;
+}
+
+@media (max-width: 900px) {
+  .section_header {
+    margin-block: 15px;
+  }
+}
+
+.category_tabs {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin: 0;
+  gap: clamp(20px, 8vw, 95px);
+}
+
+.tab {
+  background: var(--color-bg);
+  font-size: 14px;
+  cursor: pointer;
+  font-weight: 700;
+  border: 0;
+  border-bottom: 2px solid transparent;
+
+  padding: 0;
+  text-decoration: none;
+  color: inherit;
+
+  @media (min-width: 320px) {
+    font-size: 12px;
+  }
+}
+
+@media (min-width: 900px) {
+  .tab {
+    font-size: 18px;
+  }
+}
+
+.tab:hover {
+  border-bottom-color: #333;
+  color: #333;
+}
+
+.tab.active {
+  border-bottom-color: #000;
+  color: #000;
+  font-weight: 800;
+}
