+++ src/styles/mainPage/sections/reviewSection.module.css
@@ -1,0 +1,144 @@
+@import url(../../globals.css);
+@import url(../../reset/modern-reset.css);
+
+/* 리뷰 영역 */
+.section_title {
+  text-align: center;
+  font-size: 22px;
+  margin-top: 100px;
+  margin-bottom: 50px;
+  color: #333;
+  font-weight: 700;
+
+  @media (max-width: 1080px) {
+    font-size: 18px;
+    margin-top: 30px;
+    margin-bottom: 15px;
+  }
+}
+
+.review_scroll_container {
+  display: flex;
+  width: 100%;
+  scroll-behavior: smooth;
+  scrollbar-width: thin;
+  scrollbar-color: #081818 #b6c5c5;
+  cursor: grab;
+  align-self: center;
+  justify-content: center;
+  -webkit-overflow-scrolling: touch;
+  margin-bottom: 57px;
+  overflow-x: auto;
+
+  @media (min-width: 1080px) {
+    margin-bottom: 142px;
+  }
+}
+
+.review_row {
+  display: flex;
+  overflow-x: auto;
+  gap: clamp(10px, 3vw, 70px);
+  padding-bottom: clamp(30px, 5vw, 70px);
+  align-self: center;
+  justify-self: center;
+  padding-left: clamp(10px, 2vw, 20px);
+  padding-right: clamp(10px, 2vw, 20px);
+}
+
+.review_card {
+  width: clamp(100px, 28vw, 280px);
+  height: clamp(170px, 35vw, 380px);
+  color: inherit;
+  display: flex;
+  flex-shrink: 0;
+  align-items: center;
+  justify-content: center;
+}
+
+.review_main_image {
+  width: 100%;
+  height: clamp(100px, 20vw, 240px);
+  position: relative;
+  align-items: center;
+  justify-content: center;
+  margin-bottom: clamp(5px, 1vw, 10px);
+}
+
+.review_main_image img {
+  background-color: white;
+  max-width: 100%;
+  max-height: 100%;
+  object-fit: contain;
+}
+
+.review_content {
+  padding-bottom: clamp(5px, 1vw, 15px);
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  font-size: clamp(6px, 1vw, 10px);
+  width: 100%;
+}
+
+.review_title {
+  font-size: clamp(10px, 1.5vw, 16px);
+  margin-bottom: clamp(5px, 1vw, 12px);
+  color: #333;
+  font-weight: 600;
+  line-height: 1.2;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  display: -webkit-box;
+  -webkit-line-clamp: 2;
+  line-clamp: 2; /* 표준 속성 */
+  -webkit-box-orient: vertical;
+}
+
+.review_info_box {
+  display: flex;
+  gap: clamp(3px, 1vw, 10px);
+  margin-top: auto;
+}
+
+.review_thumbnail {
+  width: clamp(20px, 3vw, 40px);
+  height: clamp(20px, 3vw, 40px);
+  flex-shrink: 0;
+  overflow: hidden;
+  position: relative;
+  border-radius: 4px;
+}
+
+.review_rating {
+  color: #ffc107;
+  font-size: clamp(6px, 1vw, 10px);
+  margin-bottom: clamp(2px, 0.5vw, 5px);
+}
+
+.review_text {
+  font-size: clamp(6px, 1vw, 10px);
+  color: #666;
+  line-height: 1.2;
+  overflow: hidden;
+  white-space: nowrap;
+  text-overflow: ellipsis;
+  display: -webkit-box;
+  -webkit-line-clamp: 2;
+  line-clamp: 2;
+  -webkit-box-orient: vertical;
+  word-break: break-word;
+}
+
+@media (max-width: 480px) {
+  .review_card {
+    min-width: 90px;
+  }
+}
+
+@media (min-width: 1000px) {
+  .review_row {
+    max-width: 1000px;
+    margin: 0 auto;
+  }
+}
